rllab_implementation/requirements.txt
numpy
joblib==0.11
tensorflow-gpu==1.15.0
scipy
path
PyMC3
cached-property
pyprind
gym==0.14.0
matplotlib
